Read for full description
Removed NegativeTests test class.
All methods move to their respective pages.
