+++ .vs/config/applicationhost.config
@@ -164,7 +164,7 @@
             <site name="Blog" id="2">
                 <application path="/" applicationPool="Clr4IntegratedAppPool">
 
-                    <virtualDirectory path="/" physicalPath="C:\Users\vaneto\Documents\Visual Studio 2015\Projects\Blog-Skeleton\Blog-Skeleton\Blog" />
+                    <virtualDirectory path="/" physicalPath="E:\GitHub\FeijoaFork\Feijoa\Blog" />
 &gt;&gt;&gt;&gt;&gt;&gt;&gt; refs/remotes/origin/LectureExercises
                 </application>
                 <bindings>
