+++ Blog.UI.Tests/bin/Debug/nunit_random_seed.tmp
@@ -1,1 +1,1 @@
-1419640735
+1916031923
